+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - added hourly, daily, monthly, yearly aggregation for dashboard data
 - added support for non-git projects
 - fixed map-refresh command not to display the map
+- added OpenRouter advanced settings
 
 ## [0.24.0]